requirements.txt
ccxt==4.4.82
pandas==2.3.3
numpy==2.2.6
pandas_ta==0.4.71b0
aiogram==2.25.1
redis>=4.5.0
python-dotenv
PyYAML
loguru
httpx
# uvloop  -> ignoré sur Windows (sys_platform != "win32")
